routing: extract WireGuard TUN lifecycle from tunnel.go
Pure code-motion split of the over-threshold pkg/routing/tunnel.go
(2048 prod LOC) as partial progress on the #5661 Go control-plane
modularity cohort. The WireGuard persistent-TUN subsystem is a distinct
concern from the GRE/IPIP kernel-tunnel reconcile that dominates the
file, so it moves verbatim into a cohesive sibling.

New pkg/routing/tunnel_wireguard.go holds, byte-identical:
  - errWGIncompatibleLinkRetained sentinel (#1919 r5/r6 retry contract)
  - the WG overhead / inner-MTU constants (wgOverheadV4/V6, wgPadWorst,
    wgEngineMaxInnerMTU, wgDefaultOuterMTU) mirroring userspace-dp
  - wgTunMTUForEndpoint (#1432 S2a / #2300 / #2457 inner-MTU cap)
  - applyWireguardTunLocked (persistent TUN create/reuse + address and
    VRF reconcile, #5355 fail-closed)
  - closeTuntapFiles

No declarations renamed, no logic changed, same package so no import
cycle. tunnel.go drops from 2048 to 1795 prod LOC; its now-unused "os"
import is removed. go build ./... and go test ./pkg/routing/... pass;
gofmt clean.

Partial progress on #5661 (low-risk routing/snmp/cli files); HA-critical
files (daemon/vrrp/cluster) tracked separately.
